Add more headroom for ext4 app package containers.
Code in the volume daemon that creates the filesystem image to contain
an installed application package currently adds about 1.6% to the
size estimate that is passed down to it to account for filesystem
overhead. That figure is based upon an assumption that the filesystem
is of type FAT, but is insufficient for some big application packages
when the filesystem is of type ext4.

This change checks the filesystem type, and in the case where it's
ext4, adds about 6% to the input estimate to account for journaling,
etc.

Original commit: https://github.com/CyanogenMod/android_system_vold/commit/051ae33d3753dacc39fac87ecc300fd4908753bf

Change-Id: I823d77d440364d3d88c48d6b8bb77dc2f2941686